feat: Add pausemenu and moved options to standalone prefab
