+++ SecretHorror/Assets/Prefabs/UiPrefabs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2748d352517036c4dba2030d5d402898
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
