+++ SecretHorror/Assets/Scripts/FirstPersonCamera.cs
@@ -11,7 +11,11 @@ public class FirstPersonCamera : MonoBehaviour
     
     Camera playerCamera;
     float rotationX;
-    
+    bool freezeCamera;
+
+    public void PauseCamera(bool pause){
+            freezeCamera = pause;
+    }
     // Start is called before the first frame update
     void Start(){
         playerCamera = GetComponentInChildren<Camera>();
@@ -19,10 +23,11 @@ public class FirstPersonCamera : MonoBehaviour
 
     // Update is called once per frame
     void Update(){
-            rotationX += -Input.GetAxis("Mouse Y") * lookSpeed;
-            rotationX = Mathf.Clamp(rotationX, -lookXLimit, lookXLimit);
-            playerCamera.transform.localRotation = Quaternion.Euler(rotationX, 0, 0);
-            transform.rotation *= Quaternion.Euler(0, Input.GetAxis("Mouse X") * lookSpeed, 0);
+        if (freezeCamera) return;
+        rotationX += -Input.GetAxis("Mouse Y") * lookSpeed;
+        rotationX = Mathf.Clamp(rotationX, -lookXLimit, lookXLimit);
+        playerCamera.transform.localRotation = Quaternion.Euler(rotationX, 0, 0);
+        transform.rotation *= Quaternion.Euler(0, Input.GetAxis("Mouse X") * lookSpeed, 0);
     }
 
     void Interact(){
