+++ SecretHorror/Assets/Scripts/PlayerMovement.cs
@@ -5,25 +5,23 @@ using UnityEngine;
 [RequireComponent(typeof(CharacterController))]
 public class PlayerMovement : MonoBehaviour
 {
-    [SerializeField]
-    float walkingSpeed = 7.5f;
-    [SerializeField]
-    float runningSpeed = 11.5f;
-    [SerializeField]
-    float jumpSpeed = 8.0f;
-    [SerializeField]
-    float gravity = 20.0f;
+    [SerializeField] float walkingSpeed = 7.5f;
+    [SerializeField] float runningSpeed = 11.5f;
+    [SerializeField] float jumpSpeed = 8.0f;
+    [SerializeField] float gravity = 20.0f;
+    [SerializeField] GameObject pauseMenu;
     
     CharacterController characterController;
     Vector3 moveDirection = Vector3.zero;
+    bool paused = false;
 
     [HideInInspector]
     public bool canMove = true;
     
     // Start is called before the first frame update
     void Start(){
+        pauseMenu.SetActive(!pauseMenu.activeSelf);
         characterController = GetComponent<CharacterController>();
-
         Cursor.lockState = CursorLockMode.Locked;
         Cursor.visible = false;
         
@@ -31,6 +29,10 @@ public class PlayerMovement : MonoBehaviour
 
     // Update is called once per frame
     void Update(){
+        if (Input.GetKeyDown(KeyCode.Escape)){
+            PauseGame();
+        }
+        
         // We are grounded, so recalculate move direction based on axes
         Vector3 forward = transform.TransformDirection(Vector3.forward);
         Vector3 right = transform.TransformDirection(Vector3.right);
@@ -56,5 +58,15 @@ public class PlayerMovement : MonoBehaviour
         // Move the controller
         characterController.Move(moveDirection * Time.deltaTime);
     
+    }
+
+    void PauseGame(){
+        paused = !paused;
+        if (paused) return;
+        Time.timeScale = 0;
+        GetComponent<FirstPersonCamera>().PauseCamera(true);
+        Cursor.lockState = CursorLockMode.None;
+        Cursor.visible = true;
+        pauseMenu.SetActive(!pauseMenu.activeSelf);
     }
 }
