+++ SecretHorror/Assets/Scripts/UI/MainMenu.cs
@@ -1,57 +1,10 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using TMPro;
+using UnityEditor;
 using UnityEngine;
-using UnityEngine.Audio;
 using UnityEngine.SceneManagement;
-using UnityEngine.UI;
 
 public class MainMenu : MonoBehaviour{
-    [SerializeField] AudioMixer audioMixer;
-    [SerializeField] TMP_Dropdown resolutionDropdown;
-    [SerializeField] Slider volumeSlider;
+
     [SerializeField] GameObject optionsPanel;
-    
-    float currentVolume;
-    Resolution[] resolutions;
-
-    void Start(){
-        optionsPanel.SetActive(false);
-        resolutionDropdown.ClearOptions();
-        List<string> options = new List<string>();
-        resolutions = Screen.resolutions;
-        int currentResolutionIndex = 0;
-        
-        for (int i = 0; i < resolutions.Length; i++)
-        {
-            string option = resolutions[i].width + " x " + 
-                            resolutions[i].height;
-            options.Add(option);
-            if (resolutions[i].width == Screen.currentResolution.width 
-                && resolutions[i].height == Screen.currentResolution.height)
-                currentResolutionIndex = i;
-        }
-        
-        resolutionDropdown.AddOptions(options);
-        resolutionDropdown.RefreshShownValue();
-        LoadSettings(currentResolutionIndex);
-    }
-
-    public void SetVolume(float volume){
-        audioMixer.SetFloat("Volume", volume);
-        currentVolume = volume;
-    }
-
-    public void SetFullscreen(bool isFullscreen){
-        Screen.fullScreen = isFullscreen;
-    }
-
-    public void SetResolution(int resolutionIndex){
-        Resolution resolution = resolutions[resolutionIndex];
-        Screen.SetResolution(resolution.width, resolution.height, Screen.fullScreen);
-    }
-    
     public void OnPlayButton (){
         SceneManager.LoadScene(1);
     }
@@ -59,29 +12,7 @@ public class MainMenu : MonoBehaviour{
     public void OnQuitButton (){
         Application.Quit();
     }
-    
-    public void SaveSettings(){
-        PlayerPrefs.SetInt("ResolutionPreference", resolutionDropdown.value);
-        PlayerPrefs.SetInt("FullscreenPreference", Convert.ToInt32(Screen.fullScreen));
-        PlayerPrefs.SetFloat("VolumePreference", currentVolume); 
-    }
-    
-    public void LoadSettings(int currentResolutionIndex){
-        if (PlayerPrefs.HasKey("ResolutionPreference"))
-            resolutionDropdown.value = PlayerPrefs.GetInt("ResolutionPreference");
-        else
-            resolutionDropdown.value = currentResolutionIndex;
-        if (PlayerPrefs.HasKey("FullscreenPreference"))
-            Screen.fullScreen = Convert.ToBoolean(PlayerPrefs.GetInt("FullscreenPreference"));
-        else
-            Screen.fullScreen = true;
-        if (PlayerPrefs.HasKey("VolumePreference"))
-            volumeSlider.value = PlayerPrefs.GetFloat("VolumePreference");
-        else
-            volumeSlider.value = PlayerPrefs.GetFloat("VolumePreference");
-    }
-
     public void ToggleOptions(){
-            optionsPanel.SetActive(!optionsPanel.activeSelf);
+        optionsPanel.SetActive(!optionsPanel.activeSelf);
     }
 }
