+++ SecretHorror/Assets/Scripts/UI/OptionsMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ebf5d8fb669657a4cae12f07b35164dc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
